module: rename migrator-ee to migrator
